TECHNOLOGY WATER MANAGEMENT ON AGRICULTURE
For decades, some parts of Africa are penelised on agriculture production due to climat change that creat drouths. These phenomena increases poverty, low agriculture production, that leads to overexploitation on agriculture lands. Water has been one of the most if not all problems on agriculture, especially on some arid subshara regions. Efforts have been made through international descauses but the problem still devarstat in most or all africa regions.    

I was graduated with a Masters degree on Agriculture (Bio Agriculture) in 2018 after a professional Diploma on Telecommunications in 2015. I conducted a research that took me about 4 years to complete and to come up with a technology solution on Water Management on local agriculture for Arid Lands. I created a project called AGRO-FASO AGRIFISHFARMING.

"Agro-Faso AgriFishFarming is a community development project. With Agro-Faso in the poor rural areas will innovate many strategies to help the poor farmers by giving them training on the development of agriculture environment, water conservation, soil development, nature preservation and pollination and to reform nature to its origin to about 90%. It implicates an Android Application that helps poor farmers on how to  manage water and soil for a better agriculture production."

I do not have a degree on programming but I do have professional certificates. I manage to do programming but I have limited resources. I do not have access to app development systems. 

If you are there to make world as a better place, then I invit you to my project. I really need a team to help on creating the App.